chore: split up playwright
