+++ packages/core/src/index.ts
@@ -1,6 +1,5 @@
 export * from './events';
 export * from './autoscaling';
-export * from './browser_launchers';
 export * from './crawlers';
 export * from './enqueue_links';
 export * from './session_pool';
