+++ packages/playwright-crawler/src/internals/utils/playwright-utils.ts
@@ -21,10 +21,18 @@
 
 import ow from 'ow';
 import { Page, Response, Route } from 'playwright';
-import log from './utils_log';
-import { validators } from './validators';
-import { Request } from './request';
-import { Dictionary } from './typedefs';
+import {
+    logUtils,
+    validators,
+    Request,
+    Dictionary,
+} from '@crawlers/core';
+// import log from '../../../../core/src/utils_log';
+// import { validators } from '../../../../core/src/validators';
+// import { Request } from '../../../../core/src/request';
+// import { Dictionary } from '../../../../core/src/typedefs';
+
+const log = logUtils.child({ prefix: 'Playwright Utils' });
 
 export interface DirectNavigationOptions {
     /**
@@ -105,8 +113,3 @@ export async function gotoExtended(page: Page, request: Request, gotoOptions: Di
 
     return page.goto(url, gotoOptions);
 }
-
-/** @internal */
-export const playwrightUtils = {
-    gotoExtended,
-};
