+++ packages/puppeteer-crawler/src/index.ts
@@ -1,7 +1,23 @@
 export * from '@crawlers/browser';
 export * from './internals/puppeteer-crawler';
 export * from './internals/puppeteer-launcher';
-export * from './internals/stealth';
-export * from './internals/utils/puppeteer_request_interception';
-export * from './internals/utils/puppeteer_utils';
-export * from './internals/enqueue-links/click-elements';
+
+export * as puppeteerStealth from './internals/stealth';
+export type { StealthOptions } from './internals/stealth';
+
+export * as puppeteerRequestInterception from './internals/utils/puppeteer_request_interception';
+export type { InterceptHandler } from './internals/utils/puppeteer_request_interception';
+
+export * as puppeteerUtils from './internals/utils/puppeteer_utils';
+export type {
+    BlockRequestsOptions,
+    CompiledScriptFunction,
+    CompiledScriptParams,
+    DirectNavigationOptions,
+    InfiniteScrollOptions,
+    InjectFileOptions,
+    SaveSnapshotOptions,
+} from './internals/utils/puppeteer_utils';
+
+export * as puppeteerClickElements from './internals/enqueue-links/click-elements';
+export type { EnqueueLinksByClickingElementsOptions } from './internals/enqueue-links/click-elements';
